Clean up existing tests.
htcondor now part of stack so remove the inconsistent skipping.
Move a unit test class for an lssthtc function from
test_htcondor_services.py to test_lssthtc.py
