+++ tests/SConscript
@@ -1,3 +1,4 @@
 # -*- python -*-
 from lsst.sconsUtils import scripts
+
 scripts.BasicSConscript.tests(pyList=[])
