+++ tests/test_htcondor_service.py
@@ -28,11 +28,9 @@
 """Unit tests for the HTCondor WMS service class and related functions."""
 
 import logging
-import pathlib
-import tempfile
 import unittest
-
 import htcondor
+
 from lsst.ctrl.bps import BpsConfig, GenericWorkflowExec, GenericWorkflowJob, WmsStates
 from lsst.ctrl.bps.htcondor.htcondor_config import HTC_DEFAULTS_URI
 from lsst.ctrl.bps.htcondor.htcondor_service import (
@@ -45,7 +43,6 @@ from lsst.ctrl.bps.htcondor.htcondor_service import (
     _htc_status_to_wms_state,
     _translate_job_cmds,
 )
-from lsst.ctrl.bps.htcondor.lssthtc import _tweak_log_info
 
 logger = logging.getLogger("lsst.ctrl.bps.htcondor")
 
@@ -242,100 +239,6 @@ class HtcNodeStatusToWmsStateTestCase(unittest.TestCase):
         }
         result = _htc_node_status_to_wms_state(job)
         self.assertEqual(result, WmsStates.DELETED)
-
-
-class TweakJobInfoTestCase(unittest.TestCase):
-    """Test the function responsible for massaging job information."""
-
-    def setUp(self):
-        self.log_file = tempfile.NamedTemporaryFile(prefix="test_", suffix=".log")
-        self.log_name = pathlib.Path(self.log_file.name)
-        self.job = {
-            "Cluster": 1,
-            "Proc": 0,
-            "Iwd": str(self.log_name.parent),
-            "Owner": self.log_name.owner(),
-            "MyType": None,
-            "TerminatedNormally": True,
-        }
-
-    def tearDown(self):
-        self.log_file.close()
-
-    def testDirectAssignments(self):
-        _tweak_log_info(self.log_name, self.job)
-        self.assertEqual(self.job["ClusterId"], self.job["Cluster"])
-        self.assertEqual(self.job["ProcId"], self.job["Proc"])
-        self.assertEqual(self.job["Iwd"], str(self.log_name.parent))
-        self.assertEqual(self.job["Owner"], self.log_name.owner())
-
-    def testJobStatusAssignmentJobAbortedEvent(self):
-        job = self.job | {"MyType": "JobAbortedEvent"}
-        _tweak_log_info(self.log_name, job)
-        self.assertTrue("JobStatus" in job)
-        self.assertEqual(job["JobStatus"], htcondor.JobStatus.REMOVED)
-
-    def testJobStatusAssignmentExecuteEvent(self):
-        job = self.job | {"MyType": "ExecuteEvent"}
-        _tweak_log_info(self.log_name, job)
-        self.assertTrue("JobStatus" in job)
-        self.assertEqual(job["JobStatus"], htcondor.JobStatus.RUNNING)
-
-    def testJobStatusAssignmentSubmitEvent(self):
-        job = self.job | {"MyType": "SubmitEvent"}
-        _tweak_log_info(self.log_name, job)
-        self.assertTrue("JobStatus" in job)
-        self.assertEqual(job["JobStatus"], htcondor.JobStatus.IDLE)
-
-    def testJobStatusAssignmentJobHeldEvent(self):
-        job = self.job | {"MyType": "JobHeldEvent"}
-        _tweak_log_info(self.log_name, job)
-        self.assertTrue("JobStatus" in job)
-        self.assertEqual(job["JobStatus"], htcondor.JobStatus.HELD)
-
-    def testJobStatusAssignmentJobTerminatedEvent(self):
-        job = self.job | {"MyType": "JobTerminatedEvent"}
-        _tweak_log_info(self.log_name, job)
-        self.assertTrue("JobStatus" in job)
-        self.assertEqual(job["JobStatus"], htcondor.JobStatus.COMPLETED)
-
-    def testJobStatusAssignmentPostScriptTerminatedEvent(self):
-        job = self.job | {"MyType": "PostScriptTerminatedEvent"}
-        _tweak_log_info(self.log_name, job)
-        self.assertTrue("JobStatus" in job)
-        self.assertEqual(job["JobStatus"], htcondor.JobStatus.COMPLETED)
-
-    def testAddingExitStatusSuccess(self):
-        job = self.job | {
-            "MyType": "JobTerminatedEvent",
-            "ToE": {"ExitBySignal": False, "ExitCode": 1},
-        }
-        _tweak_log_info(self.log_name, job)
-        self.assertIn("ExitBySignal", job)
-        self.assertIs(job["ExitBySignal"], False)
-        self.assertIn("ExitCode", job)
-        self.assertEqual(job["ExitCode"], 1)
-
-    def testAddingExitStatusFailure(self):
-        job = self.job | {
-            "MyType": "JobHeldEvent",
-        }
-        with self.assertLogs(logger=logger, level="ERROR") as cm:
-            _tweak_log_info(self.log_name, job)
-        self.assertIn("Could not determine exit status", cm.output[0])
-
-    def testLoggingUnknownLogEvent(self):
-        job = self.job | {"MyType": "Foo"}
-        with self.assertLogs(logger=logger, level="DEBUG") as cm:
-            _tweak_log_info(self.log_name, job)
-        self.assertIn("Unknown log event", cm.output[1])
-
-    def testMissingKey(self):
-        job = self.job
-        del job["Cluster"]
-        with self.assertRaises(KeyError) as cm:
-            _tweak_log_info(self.log_name, job)
-        self.assertEqual(str(cm.exception), "'Cluster'")
 
 
 class HtcStatusToWmsStateTestCase(unittest.TestCase):
